test: drop PowerMock by injecting the scheduler and the clients
The two scheduler tests move into the managers' own packages so the injection
points can stay package-private.

Closes: #640
