+++ carapace-server/src/test/java/org/carapaceproxy/server/certificates/DynamicCertificatesManagerTest.java
@@ -36,6 +36,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.atLeastOnce;
@@ -56,6 +57,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import org.carapaceproxy.cluster.impl.NullGroupMembershipHandler;
@@ -66,9 +70,9 @@ import org.carapaceproxy.configstore.PropertiesConfigurationStore;
 import org.carapaceproxy.core.HttpProxyServer;
 import org.carapaceproxy.core.Listeners;
 import org.carapaceproxy.core.RuntimeServerConfiguration;
+import org.carapaceproxy.server.config.ConfigurationNotValidException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.powermock.reflect.Whitebox;
 import org.shredzone.acme4j.Certificate;
 import org.shredzone.acme4j.Login;
 import org.shredzone.acme4j.Order;
@@ -99,6 +103,48 @@ public class DynamicCertificatesManagerTest {
     }
 
     protected static final int MAX_ATTEMPTS = 7;
+
+    /**
+     * Reconfiguring the execution period from an initial value of '0' to > 0. Because of the zero period the manager
+     * never starts and when reconfigured with period > 0 it still won't run unless it was started before (#33).
+     * Relocated from ManagersExecutionTest so the scheduler injection can stay package-private.
+     */
+    @Test
+    public void testDynamicCertificatesManagerExecution() throws ConfigurationNotValidException {
+        RuntimeServerConfiguration config = new RuntimeServerConfiguration();
+        ScheduledExecutorService scheduler = mock(ScheduledExecutorService.class);
+        when(scheduler.scheduleWithFixedDelay(any(Runnable.class), anyLong(), anyLong(), any(TimeUnit.class))).thenReturn(mock(ScheduledFuture.class));
+        DynamicCertificatesManager man = new DynamicCertificatesManager(null, () -> scheduler);
+        man.setConfigurationStore(mock(ConfigurationStore.class));
+
+        // With period 0 the manager never starts
+        man.setPeriod(0);
+        man.start();
+        verify(scheduler, never()).scheduleWithFixedDelay(any(Runnable.class), anyLong(), anyLong(), any(TimeUnit.class)); // never called
+
+        // With new period >0 the manager should run whether started before
+        config.setDynamicCertificatesManagerPeriod(1);
+        man.reloadConfiguration(config);
+        assertEquals(1, man.getPeriod());
+        verify(scheduler, times(1)).scheduleWithFixedDelay(any(Runnable.class), eq(0L), eq(1L), eq(TimeUnit.SECONDS)); // once
+
+        man.stop();
+        config.setDynamicCertificatesManagerPeriod(0);
+        man.reloadConfiguration(config);
+        assertEquals(0, man.getPeriod());
+        man.start();
+        verify(scheduler, times(1)).scheduleWithFixedDelay(any(Runnable.class), eq(0L), eq(1L), eq(TimeUnit.SECONDS)); // never
+        man.stop();
+
+        // With new period >0 the manager should not run because not started before.
+        config.setDynamicCertificatesManagerPeriod(1);
+        man.reloadConfiguration(config);
+        assertEquals(1, man.getPeriod());
+        verify(scheduler, times(1)).scheduleWithFixedDelay(any(Runnable.class), eq(0L), eq(1L), eq(TimeUnit.SECONDS)); // never
+
+        man.start();
+        verify(scheduler, times(2)).scheduleWithFixedDelay(any(Runnable.class), eq(0L), eq(1L), eq(TimeUnit.SECONDS)); // once
+    }
 
     @Test
     @Parameters({
@@ -313,9 +359,9 @@ public class DynamicCertificatesManagerTest {
         }
     }
 
-    // must run after reloadConfiguration, which rebuilds the client map; by-name, because there are other map fields
+    // must run after reloadConfiguration, which rebuilds the client map
     private static void injectAcmeClients(DynamicCertificatesManager man, Map<String, ACMEClient> clients) {
-        Whitebox.setInternalState(man, "acmeClients", clients);
+        man.setAcmeClients(clients);
     }
 
     private void assertCertificateState(String domain, DynamicCertificateState expectedState, int expectedCycleCount, DynamicCertificatesManager dCMan) {
@@ -383,7 +429,7 @@ public class DynamicCertificatesManagerTest {
         when(r53Client.isDnsChallengeForDomainAvailable(any(), any())).thenReturn(
                 !(runCase.equals("challenge_creation_failed_n_reboot") || runCase.equals("challenge_check_limit_expired"))
         );
-        Whitebox.setInternalState(man, r53Client);
+        man.setRoute53Client(r53Client);
 
         // Store mocking
         ConfigurationStore store = mock(ConfigurationStore.class);
@@ -521,7 +567,7 @@ public class DynamicCertificatesManagerTest {
         when(r53Client.isDnsChallengeForDomainAvailable(any(), any())).thenReturn(
                 !(runCase.equals("challenge_creation_failed_n_reboot") || runCase.equals("challenge_check_limit_expired"))
         );
-        Whitebox.setInternalState(man, r53Client);
+        man.setRoute53Client(r53Client);
 
         // Store mocking
         ConfigurationStore store = mock(ConfigurationStore.class);
